requirements-prod.txt
azure-functions>=1.17.0
azure-identity>=1.15.0
azure-storage-blob>=12.19.0
azure-monitor-opentelemetry>=1.6.0
opentelemetry-sdk>=1.26.0
requests>=2.31.0
